Move ConcurrentEnumerator into separate file
